create WorkflowFireHistory only when Workflow fires Actions
