+++ src/sentry/workflow_engine/processors/action.py
@@ -74,37 +74,27 @@ def get_action_last_updated_statuses(now: datetime, actions: BaseQuerySet[Action
     return statuses
 
 
-def update_workflow_fire_histories(
-    actions_to_fire: BaseQuerySet[Action],
-    event_data: WorkflowEventData,
-    updates: WorkflowFireHistoryUpdates,
-) -> int:
-    # Update WorkflowFireHistory objects for workflows with actions to fire
-    fired_workflows = set(
+def create_workflow_fire_histories(
+    actions_to_fire: BaseQuerySet[Action], event_data: WorkflowEventData
+) -> list[WorkflowFireHistory]:
+    # Create WorkflowFireHistory objects for workflows we fire actions for
+    workflow_ids = set(
         WorkflowDataConditionGroup.objects.filter(
             condition_group__dataconditiongroupaction__action__in=actions_to_fire
         ).values_list("workflow_id", flat=True)
     )
 
-    logger.info(
-        "workflow_engine.workflow_fire_history.update",
-        extra={
-            "actions": [action.id for action in actions_to_fire],
-            "workflow_ids": list(fired_workflows),
-            "group_id": event_data.event.group_id,
-            "event_id": event_data.event.event_id,
-            "has_passed_filters": updates["has_passed_filters"],
-            "has_fired_actions": updates["has_fired_actions"],
-        },
-    )
-
-    updated_rows = WorkflowFireHistory.objects.filter(
-        workflow_id__in=fired_workflows,
-        group=event_data.event.group,
-        event_id=event_data.event.event_id,
-    ).update(**updates)
-
-    return updated_rows
+    fire_histories = [
+        WorkflowFireHistory(
+            workflow_id=workflow_id,
+            group=event_data.event.group,
+            event_id=event_data.event.event_id,
+            has_passed_filters=True,
+            has_fired_actions=True,
+        )
+        for workflow_id in workflow_ids
+    ]
+    return WorkflowFireHistory.objects.bulk_create(fire_histories)
 
 
 # TODO(cathy): only reinforce workflow frequency for certain issue types
@@ -116,8 +106,6 @@ def filter_recently_fired_workflow_actions(
         dataconditiongroupaction__condition_group__in=filtered_action_groups
     ).distinct()
 
-    wfh_updates = WorkflowFireHistoryUpdates(has_passed_filters=True, has_fired_actions=False)
-    update_workflow_fire_histories(actions, event_data, wfh_updates)
     group = event_data.event.group
 
     now = timezone.now()
@@ -143,8 +131,7 @@ def filter_recently_fired_workflow_actions(
     actions_without_statuses_ids = {action.id for action in actions_without_statuses}
     filtered_actions = actions.filter(id__in=actions_to_include | actions_without_statuses_ids)
 
-    wfh_updates["has_fired_actions"] = True
-    update_workflow_fire_histories(filtered_actions, event_data, wfh_updates)
+    create_workflow_fire_histories(filtered_actions, event_data)
 
     return filtered_actions
