+++ src/sentry/workflow_engine/processors/workflow.py
@@ -17,7 +17,6 @@ from sentry.workflow_engine.models import (
     DataConditionGroup,
     Detector,
     Workflow,
-    WorkflowFireHistory,
 )
 from sentry.workflow_engine.processors.action import filter_recently_fired_workflow_actions
 from sentry.workflow_engine.processors.data_condition_group import process_data_condition_group
@@ -58,19 +57,6 @@ def delete_workflow(workflow: Workflow) -> bool:
         workflow.delete()
 
     return True
-
-
-def create_workflow_fire_histories(
-    workflows: set[Workflow], event_data: WorkflowEventData
-) -> list[WorkflowFireHistory]:
-    # Create WorkflowFireHistory objects for triggered workflows
-    fire_histories = [
-        WorkflowFireHistory(
-            workflow=workflow, group=event_data.event.group, event_id=event_data.event.event_id
-        )
-        for workflow in workflows
-    ]
-    return WorkflowFireHistory.objects.bulk_create(fire_histories)
 
 
 def enqueue_workflow(
@@ -125,8 +111,6 @@ def evaluate_workflow_triggers(
         else:
             if evaluation:
                 triggered_workflows.add(workflow)
-
-    create_workflow_fire_histories(triggered_workflows, event_data)
 
     return triggered_workflows
