+++ tests/sentry/workflow_engine/processors/test_workflow.py
@@ -26,7 +26,6 @@ from sentry.workflow_engine.models.data_condition import Condition
 from sentry.workflow_engine.processors.workflow import (
     WORKFLOW_ENGINE_BUFFER_LIST_KEY,
     WorkflowDataConditionGroupType,
-    create_workflow_fire_histories,
     delete_workflow,
     enqueue_workflow,
     evaluate_workflow_triggers,
@@ -584,37 +583,7 @@ class TestEvaluateWorkflowActionFilters(BaseWorkflowTest):
 
         # TODO @saponifi3d - Add a check to ensure the second condition is enqueued for later evaluation
 
-
-class TestWorkflowFireHistory(BaseWorkflowTest):
-    def setUp(self):
-        (
-            self.workflow,
-            self.detector,
-            self.detector_workflow,
-            self.workflow_triggers,
-        ) = self.create_detector_and_workflow()
-
-        self.action_group, self.action = self.create_workflow_action(workflow=self.workflow)
-
-        self.group, self.event, self.group_event = self.create_group_event(
-            occurrence=self.build_occurrence(evidence_data={"detector_id": self.detector.id})
-        )
-        self.event_data = WorkflowEventData(event=self.group_event)
-
-    def test_create_workflow_fire_histories(self):
-        create_workflow_fire_histories({self.workflow}, self.event_data)
-        assert (
-            WorkflowFireHistory.objects.filter(
-                workflow=self.workflow,
-                group=self.group,
-                event_id=self.group_event.event_id,
-                has_fired_actions=False,
-            ).count()
-            == 1
-        )
-
-    def test_evaluate_filters_updates_histories(self):
-        # only updates workflows that meet filters
+    def test_creates_histories(self):
         self.create_data_condition(
             condition_group=self.action_group,
             type=Condition.EVENT_SEEN_COUNT,
@@ -631,28 +600,18 @@ class TestWorkflowFireHistory(BaseWorkflowTest):
             condition_result=True,
         )  # condition is not met
 
-        create_workflow_fire_histories({self.workflow, workflow}, self.event_data)
-
         triggered_actions = evaluate_workflows_action_filters(
             {self.workflow, workflow}, self.event_data
         )
         assert set(triggered_actions) == {self.action}
 
+        assert WorkflowFireHistory.objects.all().count() == 1
         assert (
             WorkflowFireHistory.objects.filter(
                 workflow=self.workflow,
                 group=self.group,
                 event_id=self.group_event.event_id,
                 has_fired_actions=True,
-            ).count()
-            == 1
-        )
-        assert (
-            WorkflowFireHistory.objects.filter(
-                workflow=workflow,
-                group=self.group,
-                event_id=self.group_event.event_id,
-                has_fired_actions=False,
             ).count()
             == 1
         )
